Former-commit-id: c5620e9fdef1893af0c175b68595ee53de244e80

+++ Assets/Engine/Shaders/UnderWater.shader
@@ -53,7 +53,7 @@ Shader "Hungry Dragon/UnderWater"
 
 				struct v2f {
 					float4 vertex : SV_POSITION;
-					half2 uv : TEXCOORD0;
+					float2 uv : TEXCOORD0;
 					float4 scrPos:TEXCOORD1;
 					float4 color : COLOR;
 				};
@@ -77,8 +77,6 @@ Shader "Hungry Dragon/UnderWater"
 					v.vertex.y += (sinX + sinY) * 0.15 * moveVertex * v.color.w;
 
 					o.vertex = UnityObjectToClipPos(v.vertex);
-
-
 					o.scrPos = ComputeScreenPos(o.vertex);
 					o.uv = TRANSFORM_TEX(v.uv, _MainTex);
